% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_html}
\alias{as_html}
\title{Converts a Markdown string into HTML}
\usage{
as_html(x)
}
\arguments{
\item{x}{A Markdown-formatted character string.}
}
\value{
A HTML-formatted character string.
}
\description{
Converts a Markdown string into HTML
}
