% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modules_sets.R
\name{get_modules_sets}
\alias{get_modules_sets}
\title{Get All Environment Modules in a Module Path}
\usage{
get_modules_sets(module_repositories, ...)
}
\arguments{
\item{module_repositories}{...}

\item{\dots}{Additional arguments passed to \code{\link[=module_avail]{module_avail()}}.}
}
\value{
A list ...
}
\description{
Get All Environment Modules in a Module Path
}
