% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_path.R
\name{module_path}
\alias{module_path}
\alias{MODULEPATH}
\title{Get the module path}
\usage{
module_path()
}
\value{
A character vector of environment module folders.
}
\description{
Parses environment variable \code{MODULEPATH}.
}
