% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_module.R
\name{parse_module}
\alias{parse_module}
\title{Parses one Module Environment}
\usage{
parse_module(m)
}
\arguments{
\item{m}{A list ...}
}
\value{
A list of length \code{length(x)} ...
}
\description{
Parses one Module Environment
}
