% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_widgets.R
\name{repos_buttons}
\alias{repos_buttons}
\title{Output HTML JQuery Navigation Buttons}
\usage{
repos_buttons(modules_sets)
}
\arguments{
\item{modules_sets}{A named list ...}
}
\value{
Nothing.
}
\description{
Output HTML JQuery Navigation Buttons
}
