% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_avail.R
\name{module_avail}
\alias{module_avail}
\title{Get All Environment Modules in a Module Path}
\usage{
module_avail(
  info,
  distros = character(0L),
  onMissingPath = getOption("onMissingPath", c("okay", "error", "warning", "ignore")),
  force = FALSE
)
}
\arguments{
\item{info}{...}

\item{distros}{(optional) Character vector of Linux-distribution specific software to list.}

\item{onMissingPath}{...}

\item{force}{If TRUE, cached results are ignored.}
}
\value{
...
}
\description{
Get All Environment Modules in a Module Path
}
