% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spider.R
\name{spider}
\alias{spider}
\title{Module Spider on a Module Path}
\usage{
spider(module_path, force = FALSE)
}
\arguments{
\item{module_path}{A folder with environment modules.}

\item{force}{If FALSE, and results are already cached, then the
cached results are returned.}
}
\value{
A non-parsed, JSON string.
}
\description{
Module Spider on a Module Path
}
